+++ src/components/BookCardDetails/BookCardDetails.css
