+++ src/components/SearchPanel/SearchPanel.css
