+++ src/components/SearchPanel/SearchPanel.tsx
