+++ src/pages/HomePage/HomePage.tsx
@@ -1,4 +1,4 @@
-import SearchPanel from './components/SearchPanel/SearchPanel';
+import SearchPanel from '../../components/SearchPanel/SearchPanel';
 import './HomePage.css';
 
 const HomePage = () => {
